Update merge file to skip build if more conflicts

+++ merge.sh
@@ -10,6 +10,13 @@ git checkout -- locales/es_ES/LC_MESSAGES/messages.po
 echo "Merging from master"
 git fetch
 git merge origin/master
+echo "Are there any conflicts (not the po file)? [y/N]"
+read input
+if [ input == 'y' ]; then 
+  abort=1
+else
+  abort=0
+fi
 echo "Merging pofile using msgcat"
 msgcat --use-first messages.local.po messages.remote.po -o messages.merge.po
 mv messages.merge.po locales/es_ES/LC_MESSAGES/messages.po
@@ -18,21 +25,23 @@ rm messages.local.po
 rm messages.remote.po
 echo "Updating python versions"
 pip install -r requirements.txt
-echo "Updating pot file to check strings after merge..."
-if [ -f locales/messages.pot ];
-then
-echo "Erasing old pot file (local cache)..."
-rm locales/messages.pot
-fi
-echo "Building docs to get new pot file..."
-mkdocs build -f mkdocs_es.yml --clean
-echo "Merging pofiles..."
-msgmerge -U locales/es_ES/LC_MESSAGES/messages.po locales/messages.pot
-git add locales/es_ES/LC_MESSAGES/messages.po
-if python check_strings.py;
-then
-  echo "DONE - remember to 'git commit' to end merging"
-else
-  echo "Test failed, update the strings with poedit"
-  echo "DONE - use 'git commit' to end merging."
+if [ abort == 0 ]; then
+  echo "Updating pot file to check strings after merge..."
+  if [ -f locales/messages.pot ];
+  then
+    echo "Erasing old pot file (local cache)..."
+    rm locales/messages.pot
+  fi
+  echo "Building docs to get new pot file..."
+  mkdocs build -f mkdocs_es.yml --clean
+  echo "Merging pofiles..."
+  msgmerge -U locales/es_ES/LC_MESSAGES/messages.po locales/messages.pot
+  git add locales/es_ES/LC_MESSAGES/messages.po
+  if python check_strings.py;
+  then
+    echo "DONE - remember to 'git commit' to end merging"
+  else
+    echo "Test failed, update the strings with poedit"
+    echo "DONE - use 'git commit' to end merging."
+  fi
 fi
